Added envir to do.call in +.mutable
There was an issue with hlf mutableResponse concerning the
summaryColumn being passed as a symbol to mutable.default.  The +
function could therefore not find it, since that variable was only
bound in its calling environment.  Have to think carefully about this
since a call to parent.frame is involved.  May have to revert.